install dependencies into separate container
that way, the final container will be much smaller and faster to build. Enabling rebuilding the container on every repository push. We could then also run the tests inside the container if we want.

The deps container only needs to be rebuilt if the dependencies changed. We should also rebuild the deps container for every release, so that the dependencies in the release container are up to date.